Extract helper function `find_free_net_port` to golem.utils.
